Memfault Firmware SDK 1.22.0 (Build 13216)

+++ components/demo/src/panics/memfault_demo_panics.c
@@ -223,7 +223,7 @@ int memfault_demo_cli_loadaddr(int argc, char *argv[]) {
     return -1;
   }
   uint32_t addr = (uint32_t)strtoul(argv[1], NULL, 0);
-  uint32_t val = *(uint32_t *)addr;
+  uint32_t val = *(uint32_t *)(uintptr_t)addr;
 
   MEMFAULT_LOG_INFO("Read 0x%08" PRIx32 " from 0x%08" PRIx32, val, (uint32_t)(uintptr_t)addr);
   return 0;
